feat: use pip0001 prefix for provider discovery event types
