+++ packages/light-client-extension-helpers/src/web-page/web-page-helper.ts
@@ -13,7 +13,7 @@ import {
 import type { LightClientProvider, RawChain } from "./types"
 
 export type * from "./types"
-export type * from "./PIP6963"
+export type * from "./PIP0001"
 
 const postToExtension = (message: PostMessage<ToExtension>) =>
   window.postMessage(message, window.origin)
